requirements.txt
flask-sqlalchemy
sqlalchemy_utils
flask-mysql
flask-migrate
mysqlclient
Flask-Security
oauth2client
google-api-python-client
facebook-sdk
bcrypt
flask-dance
pyopenssl
Markdown
flask_restful